requirements.txt
playwright
ddddocr
python-dotenv
schedule
